move fixture page detection to commercial features constructor
